sensor-server/pyproject.toml
[tool.poetry]
name = "sensor-server"
version = "0.1.0"
description = ""
authors = [
    "Łukasz Suchołbiak <sucholbiaklukasz@gmail.com>",
    "Bartosz Nowak <nowobart@gmail.com>",
    "Michał Pałasz <michal.palasz12@gmail.com>",
    "Maciej Scheffer <placeholder@example.com>"
]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
python-mbedtls = "^2.10.1"
cryptography = "^44.0.0"
python3-dtls = "^1.3.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.0"
black = "^24.10.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
